create: use hill radius to compute default ring gap
Note that the computed distances between rings are false, since the hill
radius of planets of a ring are greater than hill radius of planets in
previously generated ring.
As the second only is used to compute the distances between rings, the
final ring gaps are all false.

A out->int generation would fix the problem *if all planets have the
same mass*, so the best way to handle that is to properly compute a
distance between ring so that no planet is in hill sphere of another.